Promote KernelCare if not installed.
Case SWAT-22, SWAT-320: Added a check to see if KC is installed, if not
provide an "info" advisory with a link to the KernelCare site for
buying/installing.

The decision whether or not to show the info advisory is based on three
criteria:
* If KC is already installed, don't bother.
* If we're in an environment that *can't* use KC, don't bother.
* If the partner for this server has decreed that "thou shalt not,"
  then don't. If they've listed an alternative path for that--via
  another website or email address, show it, but use that link instead.

I can't take credit for all of this--some of it was Brett's original
info banner implementation.